minor fixes to use floats of model draw functions
git-svn-id: https://svn.csail.mit.edu/drc@1045 c7283977-0100-402a-a91a-fa70b306dbfe

+++ software/utils/visualization_utils/src/InteractableGlKinematicBody.hpp
@@ -48,7 +48,7 @@ class InteractableGlKinematicBody: public GlKinematicBody
   
    // double c[3] = {0.3,0.3,0.3};
    // double alpha = self->alpha;
-   void draw_body (double (&c)[3], double alpha)
+   void draw_body (float (&c)[3], float alpha)
    {
      
       glColor4f(c[0],c[1],c[2],alpha);
